backend/requirements.txt
# backend/requirements.txt
flask
flask_cors
tensorflow
scikit-learn
pandas
numpy
python-dotenv
requests
google-generativeai
dateparser